refactor(resources): keep readSocket internal
`resources/index.ts` does `export * from './realtime'`, so putting the helper
there would have made it a public root export — frozen at 1.0.0 under ADR-0010
and the at-bar rule. It is a workaround, not an API.
